+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/markers/ProofMarkers.kt
@@ -120,26 +120,3 @@ fun CoercionProof.coercionMessage(): String =
     Link to proof declaration:
     <code lang="kotlin">$through</code>
   """.trimIndent()
-
-/**
- * Similar to [arrow.meta.ide.dsl.editor.lineMarker.LineMarkerSyntax.addLineMarkerProvider], is an extension for
- * PsiElements that are not leafs so it will look for the first Leaf corresponding the targeted psiElement
- */
-@Suppress("UNCHECKED_CAST")
-fun <A : PsiElement> IdeMetaPlugin.addLineMarkerProvider(
-  icon: Icon,
-  transform: (PsiElement) -> A?,
-  composite: Class<A>,
-  message: DescriptorRenderer.Companion.(A) -> String = Noop.string2(),
-  placed: GutterIconRenderer.Alignment = GutterIconRenderer.Alignment.RIGHT,
-  navigate: (event: MouseEvent, element: PsiElement) -> Unit = Noop.effect2,
-  clickAction: AnAction? = null
-): ExtensionPhase =
-  addLineMarkerProvider(
-    { PsiTreeUtil.findChildOfType(transform(it), LeafPsiElement::class.java) },
-    {
-      it.onComposite(composite) { psi: A ->
-        lineMarkerInfo(icon, it, { message(DescriptorRenderer.Companion, psi) }, placed, navigate, clickAction)
-      }
-    }
-  )
