+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/proofs/markers/valueArgumentCoercion.kt
@@ -32,27 +32,3 @@ val IdeMetaPlugin.implicitCoercionValueArgumentLineMarker: ExtensionPhase
       } ?: "Proof not found"
     }
   )
-
-/**
- * Similar to [arrow.meta.ide.dsl.editor.lineMarker.LineMarkerSyntax.addLineMarkerProviderM], is an extension for
- * PsiElements that are not leafs so it will look for the first Leaf corresponding the targeted psiElement
- */
-internal fun <A : PsiElement> IdeMetaPlugin.addLineMarkerProviderM(
-  icon: Icon,
-  transform: (PsiElement) -> A?,
-  composite: Class<A>,
-  message: DescriptorRenderer.Companion.(A) -> String = Noop.string2(),
-  commonIcon: MergeableLineMarkerInfo<PsiElement>.(others: List<MergeableLineMarkerInfo<PsiElement>>) -> Icon = { icon },
-  mergeWith: MergeableLineMarkerInfo<PsiElement>.(other: MergeableLineMarkerInfo<*>) -> Boolean = { this.icon == it.icon },
-  navigate: (event: MouseEvent, element: PsiElement) -> Unit = Noop.effect2,
-  placed: GutterIconRenderer.Alignment = GutterIconRenderer.Alignment.RIGHT,
-  clickAction: AnAction? = null
-): ExtensionPhase =
-  addLineMarkerProvider(
-    { PsiTreeUtil.findChildOfType(transform(it), LeafPsiElement::class.java) },
-    {
-      it.onComposite(composite) { psi: A ->
-        mergeableLineMarkerInfo(icon, it, { message(DescriptorRenderer.Companion, psi) }, commonIcon, mergeWith, placed, navigate, clickAction)
-      }
-    }
-  )
